🍒 Manual backport - First week of month not being visible in calendar selector (#29728)
